+++ coprocessorInt/src/LMS7002M_lib/LMS7002M_filter_cal.c
@@ -58,9 +58,9 @@ int cal_gain_selection(LMS7002M_t *self, const LMS7002M_chan_t channel)
     return cal_read_rssi(self, channel);
 }
 
-int cal_setup_cgen(LMS7002M_t *self, const double bw)
+int cal_setup_cgen(LMS7002M_t *self, const float bw)
 {
-    double cgen_freq = bw*20;
+    float cgen_freq = bw*20;
     if (cgen_freq < 60e6) cgen_freq = 60e6;
     if (cgen_freq > 640e6) cgen_freq = 640e6;
     while ((int)(cgen_freq/1e6) == (int)(bw/16e6)) cgen_freq -= 10e6;
